Code refactoring: 1. For all measuredtools add superclass measuredtool with common methods: isMultiTool()=true, enable(), disable(), measured(), measuredClose() 2. Add method isMultiTool()=false to superclass maptool 3. Set singleton true for all maptools 4. Remove activeToolName in components/flexberry-maptoolbar.js, rewrite code using isMultiTool method
