+++ addon/maptools/markermeasuretool.js
@@ -1,37 +1,10 @@
-import MapTool from 'ember-flexberry-gis/maptools/maptool';
+import MeasureTool from 'ember-flexberry-gis/maptools/measuretool';
 
-export default MapTool.extend({
-
-  _layer: null,
-  _measure: null,
-
-  markerClose(e) {
-    this.markerLayer.disableEdit();
-    this.get('map').off('measure:created');
-    this.markerLayer.remove();
-  },
-
-  markerMeasured(e) {
-    this.markerLayer = e.layer;
-    this.get('map').off('measure:created');
-  },
+export default MeasureTool.extend({
 
   enable() {
     this._super(...arguments);
-    // there are problems if dragging is disabled
-    var map = this.get('map');
-    map.dragging.enable();
-    this._measure = map.MeasureTools ? map.MeasureTools : L.measureBase(map);
     this._measure.markerBaseTool.startMeasure();
-    this.get('map').on('measure:created', this.markerMeasured, this);
-  },
-
-  disable() {
-    this._super(...arguments);
-    var map = this.get('map');
-    map.dragging.disable();
-    this.markerClose();
-    map.editTools.stopDrawing();
-  },
+  }
 
 });
