+++ addon/maptools/polylinemeasuretool.js
@@ -1,36 +1,10 @@
-import MapTool from 'ember-flexberry-gis/maptools/maptool';
+import MeasureTool from 'ember-flexberry-gis/maptools/measuretool';
 
-export default MapTool.extend({
-
-  _layer: null,
-  _measure: null,
-
-  polylineClose(e) {
-    this.polylineLayer.disableEdit();
-    this.get('map').off('measure:created');
-    this.polylineLayer.remove();
-  },
-
-  polylineMeasured(e) {
-    this.polylineLayer = e.layer;
-    this.get('map').off('measure:created');
-  },
+export default MeasureTool.extend({
 
   enable() {
     this._super(...arguments);
-    // there are problems if dragging is disabled
-    var map = this.get('map');
-    map.dragging.enable();
-    this._measure = map.MeasureTools ? map.MeasureTools : L.measureBase(map);
     this._measure.polylineBaseTool.startMeasure();
-    this.get('map').on('measure:created', this.polylineMeasured, this);
-  },
+  }
 
-  disable() {
-    this._super(...arguments);
-    var map = this.get('map');
-    map.dragging.disable();
-    this.polylineClose();
-    map.editTools.stopDrawing();
-  },
 });
