+++ app/maptools/measuretool.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-flexberry-gis/maptools/measuretool';
